% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gamma_terms.R
\name{get_gamma_terms}
\alias{get_gamma_terms}
\title{get gamma terms object}
\usage{
get_gamma_terms(top_n_beta_obj, td_gamma_obj, criteria = "frequency", topicModel = NULL)
}
\arguments{
\item{top_n_beta_obj}{can be created with topicHelpers::get_top_n_beta_terms}

\item{td_gamma_obj}{can be created by topicHelpers::get_td_gamma}

\item{criteria}{one of either "frequency" for highest frequency words or "frex" for the "weighted harmonic mean of words'
ranks in terms of exclusivity and frequency" (Roberts, Stewart, Tingley)
(See https://cran.r-project.org/web/packages/stm/vignettes/stmVignette.pdf).}

\item{topicModel}{if criteria = "frex", this is the stm topic model object from which to extract the "frex" indices.
Defaults to highest frequency words}
}
\value{
a tbl_df containing the top n gamma terms for each topic
}
\description{
get gamma terms object from an stm topic model object and its top_n_beta_terms object
}
\examples{
\dontrun{
data("tm_ex_dat")

td_beta <- get_td_beta(tm_ex_dat$topic_model_test)

top_n_beta <- get_top_n_beta_terms(td_beta)

td_gamma <- get_td_gamma(tm_ex_dat$topic_model_test)

get_gamma_terms(top_n_beta, td_gamma)

get_gamma_terms(top_n_beta, td_gamma, criteria = "frex", topicModel = tm_ex_dat$topic_model_test)
}
}
