% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_td_beta.R
\name{get_td_beta}
\alias{get_td_beta}
\title{get the beta matrix from an stm topic model object}
\usage{
get_td_beta(topicModel)
}
\arguments{
\item{topicModel}{the stm topic model from which to extract the beta matrix}
}
\value{
a tbl_df with columns "topic", "term", and "beta"
}
\description{
get a broom::tidy matrix containing the probabiliies each word is generated from each topic
}
\examples{
\dontrun{
data("tm_ex_dat")

get_td_beta(tm_ex_dat$topic_model_test)
}
}
