% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_top_n_beta_terms.R
\name{get_top_n_beta_terms}
\alias{get_top_n_beta_terms}
\title{get top n beta terms for each topic in an stm topic model}
\usage{
get_top_n_beta_terms(td_beta_obj, n = 7)
}
\arguments{
\item{td_beta_obj}{a tidied beta matrix from the stm topic model of interest. Can be created with topicHelpers::get_td_beta}

\item{n}{the number of top terms to be returned for each topic}
}
\value{
a tbl_df with columns "topic" and "terms"
}
\description{
get the top n beta terms for each topic in an stm topic model
}
\examples{
\dontrun{
data("tm_ex_dat")

beta_obj <- get_td_beta(tm_ex_dat$topic_model_test)

get_top_n_beta_terms(beta_obj)
}
}
