% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_topics_by_terms.R
\name{get_topics_by_terms}
\alias{get_topics_by_terms}
\title{search gamma_terms for words}
\usage{
get_topics_by_terms(gamma_terms, search_words, terms_col, n = 7, ...)
}
\arguments{
\item{gamma_terms}{a tibble or dataframe containing terms by gamma by topic}

\item{search_words}{a character string or vector of the word or words to search for}

\item{terms_col}{the column containing the terms, which will be separated into n columns}

\item{n}{the number of words to keep in for each topic,}

\item{...}{additional arguments passed}
}
\value{
a tibble with columns "topic", "gamma", "term_1", ..., "term_n"
}
\description{
search a gamma_terms object for a speicific word or two
}
\note{
Columns term_1, term_2, ..., term_n will be created by the function, 
as terms_col is separated to make a separate column for each value in 1 to n
}
\examples{
\dontrun{
data("tm_ex_dat")

td_beta <- get_td_beta(tm_ex_dat$topic_model_test)

top_n_beta <- get_top_n_beta_terms(td_beta)

td_gamma <- get_td_gamma(tm_ex_dat$topic_model_test)

my_gamma_terms <- get_gamma_terms(top_n_beta, td_gamma)

get_topics_by_terms(my_gamma_terms, search_words = c("government", "taliban"), terms)
}
}
