% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_tidy_text_stm_ready.R
\name{make_tidy_text_stm_ready}
\alias{make_tidy_text_stm_ready}
\title{make an stm::prepDocuments() object ready for stm::stm}
\usage{
make_tidy_text_stm_ready(tidy_train_text_obj)
}
\arguments{
\item{tidy_train_text_obj}{the "tidy text" object that is to be processed. see https://juliasilge.com/blog/evaluating-stm/}
}
\value{
a list object containing "documents", "vocab", "meta", etc,. typically returned by stm::prepDocuments()
}
\description{
get a pre-processed stm::prepDocuments() object ready for use in stm::stm(). This function allows for the
expedition of preparing a tidy text object for use in stm::stm()
}
\examples{
\dontrun{
data("tm_ex_dat")

tidy_train_text <- tm_ex_dat$tidy_train_text

nested_dat <- tm_ex_dat$nested_tidy_train_text

make_tidy_text_stm_ready(tidy_train_text)
}
}
