% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_topic_prevalence_plot.R
\name{make_topic_prevalence_plot}
\alias{make_topic_prevalence_plot}
\title{make a plot showing the top n most prevalent topics with associated words}
\usage{
make_topic_prevalence_plot(gamma_terms_obj, k = 20)
}
\arguments{
\item{gamma_terms_obj}{can be created by topicHelpers::get_gamma_terms}

\item{k}{the number of topics to plot}
}
\description{
make a plot showing the top n most prevalent topics with associated gamma terms
}
\note{
additional ggplot2 elements can be added, i.e  make_topic_prevalence_plot(gamma_terms_object) + labs(title = "title")
}
\examples{
\dontrun{
data("tm_ex_dat")

td_beta <- get_td_beta(tm_ex_dat$topic_model_test)

top_n_beta <- get_top_n_beta_terms(td_beta)

td_gamma <- get_td_gamma(tm_ex_dat$topic_model_test)

my_gamma_terms <- get_gamma_terms(top_n_beta, td_gamma)

make_topic_prevalence_plot(my_gamma_terms)

make_topic_prevalence_plot(my_gamma_terms, k = 50)
}
}
