% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort_all_document_topics.R
\name{sort_all_document_topics}
\alias{sort_all_document_topics}
\title{get all topic importance rankings for every document}
\usage{
sort_all_document_topics(data, id_column, n_topics, ...)
}
\arguments{
\item{data}{df or tibble containing betas by topic for one document}

\item{id_column}{the column containing the unique integer id of the row to be sorted}

\item{n_topics}{the number of topics to keep for each document}

\item{...}{additional arguments passed}
}
\description{
sort topic importance for each document (row)
}
\note{
to be used within topicHelpers::make_td_thetas_df. Shouldn't need to be used outside of that context
}
