% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort_document_topics.R
\name{sort_document_topics}
\alias{sort_document_topics}
\title{sort topics in a row by beta}
\usage{
sort_document_topics(data, id_column, sort_id, n_topics, values = TRUE, ...)
}
\arguments{
\item{data}{df or tibble containing betas by topic for one document}

\item{id_column}{the column containing the unique id of the row to be sorted}

\item{sort_id}{the unique integer id of the row, from id_column, to be sorted}

\item{n_topics}{the number of topics to keep}

\item{values}{should the topic probabilities be returned (TRUE), or only the topic names (FALSE)}

\item{...}{additional arguments passed}
}
\description{
filter for a row and sort it by topic importance
}
