% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ex_data.R
\docType{data}
\name{tm_ex_dat}
\alias{tm_ex_dat}
\title{Example data from a simple stm topic model, and it's required pre-processing steps.
objects include:
train_text - original data frame
tidy_train_text - words unnested and other filtering performed
nested_tidy_train_text - nested tidy_train_text
train_text_prepped - result of stm::prepDocuments
topic_model_test - result of stm::stm, has no content and no province arguments
topic_model_test_2 - result of stm::stm, has no content and no province arguments}
\format{
an object of class RData
}
\usage{
data(tm_ex_dat)
}
\description{
Example data from a simple stm topic model, and it's required pre-processing steps.
objects include:
train_text - original data frame
tidy_train_text - words unnested and other filtering performed
nested_tidy_train_text - nested tidy_train_text
train_text_prepped - result of stm::prepDocuments
topic_model_test - result of stm::stm, has no content and no province arguments
topic_model_test_2 - result of stm::stm, has no content and no province arguments
}
\examples{
data(tm_ex_dat)

}
\keyword{datasets}
